+++ libkwave/MimeData.cpp
@@ -263,7 +263,8 @@ sample_index_t Kwave::MimeData::decode(QWidget *widget, const QMimeData *e,
 	// if the sample rate has to be converted, adjust the length
 	// right border
 	if (!qFuzzyCompare(src_rate, dst_rate) && (dst_rate > 1) && sig.tracks())
-	    decoded_length = qRound(decoded_length * (dst_rate / src_rate));
+	    decoded_length = qRound(static_cast<double>(decoded_length) *
+	        (dst_rate / src_rate));
 
 	sample_index_t left  = pos;
 	sample_index_t right = left + decoded_length - 1;
